requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
requests>=2.25.0
flask-cors==3.0.10
python-dotenv==1.0.1
httpx==0.24.1
ffmpeg-python
youtube_dl
yt-dlp